OP-1444: Block saving/editing historical payer
